Revert "!65626 Update cann or other packages"
This reverts commit 3fa1cdb2c543c55a7f30e510581c221943ed0d0d, reversing
changes made to a35cdc15084b1ba62f30349410c7e58e944ea1db.

+++ mindspore/python/mindspore/mindrecord/tools/cifar100_to_mr.py
@@ -18,6 +18,7 @@ Cifar100 convert tool for MindRecord.
 
 from importlib import import_module
 import os
+import numpy as np
 
 from mindspore import log as logger
 from .cifar100 import Cifar100
@@ -147,8 +148,8 @@ class Cifar100ToMR:
         """
         raw_data = []
         for i, img in enumerate(images):
-            fine_label = fine_labels[i][0]
-            coarse_label = coarse_labels[i][0]
+            fine_label = np.int(fine_labels[i][0])
+            coarse_label = np.int(coarse_labels[i][0])
             _, img = self.cv_import.imencode(".jpeg", img[..., [2, 1, 0]])
             row_data = {"id": int(i),
                         "data": img.tobytes(),
